Refactoring to use the new Drupal console extend functionality
