+++ .gitignore
@@ -1,0 +1,6 @@
+# OS, IDE
+.DS_Store
+.idea
+
+# Composer
+/vendor
