--- chain.yml
@@ -1,18 +1,0 @@
-# Custom chains
-chain:
-  name:
-    'site:new':
-      file: '~/.config/composer/dev_scripts/vendor/dennisdigital/drupal_console_commands/config/chain/chain-site-new.yml'
-    'site:build':
-      file: '~/.config/composer/dev_scripts/vendor/dennisdigital/drupal_console_commands/config/chain/chain-site-build.yml'
-    'site:install':
-      file: '~/.config/composer/dev_scripts/vendor/dennisdigital/drupal_console_commands/config/chain/chain-site-install.yml'
-    'site:npm':
-      file: '~/.config/composer/dev_scripts/vendor/dennisdigital/drupal_console_commands/config/chain/chain-npm.yml'
-    'site:grunt':
-      file: '~/.config/composer/dev_scripts/vendor/dennisdigital/drupal_console_commands/config/chain/chain-grunt.yml'
-  disabled:
-    'test:main':
-      file: '~/.config/composer/dev_scripts/vendor/dennisdigital/drupal_console_commands/config/chain/chain-test-main.yml'
-    'test:child':
-      file: '~/.config/composer/dev_scripts/vendor/dennisdigital/drupal_console_commands/config/chain/chain-test-child.yml'
