+++ chain/chain-sn.yml
@@ -1,0 +1,8 @@
+commands:
+# Build from template
+  - command: chain:site:new
+    options:
+      placeholder:
+        - 'project:balblabal'
+        - 'version:2'
+        - 'directory:/'
